move components into src/components
